Single excitation
The initial fluctuations of the Choi's purity describes the time it takes a random single qubit state placed at the end of the environment chain to "communicate" with the initial qubit state of the chaometer. As the chain size increases, it takes more time to the fluctuation to appear, is this the famous "back flow of information"?